Wrap the whole app in the DicePool prototype switcher
DicePool itself now reads the nearest Prototype selection via a new
usePrototypeVersion() hook and swaps in the matching variant, falling
through to today's design whenever there's no ancestor Prototype (every
existing test). __root.tsx wraps the entire app in one Prototype, so the
switcher bar and every real DicePool usage — the Defense tab's resistance
pools, the attack/defense calculators, the legwork dialog, spirit
summoning, and the demo pools — change together across the whole app
instead of just the Defense-tab demo section.

Documents the pattern (and the deep-nesting-across-the-whole-app case that
motivated it) in docs/ui/prototype.md.
