+++ src/components/system/dicePool/prototype/dicePoolPrototypeVersions.ts
@@ -1,0 +1,10 @@
+import type { PrototypeVersion } from "#/components/ui/prototype/prototype.tsx"
+
+export const dicePoolPrototypeVersions: PrototypeVersion[] = [
+  { key: "current", name: "Current design" },
+  { key: "A", name: "A — Ledger + Roll Footer" },
+  { key: "B", name: "B — Hero Dial + Chips" },
+  { key: "C", name: "C — Compact Row" },
+  { key: "D", name: "D — Split Panel" },
+  { key: "E", name: "E — Minimal HUD Pill" },
+]
