+++ src/routes/__root.tsx
@@ -4,8 +4,10 @@ import Stack from "@mui/material/Stack"
 import { TanStackDevtools } from "@tanstack/react-devtools"
 import { createRootRouteWithContext, Outlet } from "@tanstack/react-router"
 
+import { dicePoolPrototypeVersions } from "#/components/system/dicePool/prototype/dicePoolPrototypeVersions.ts"
 import Footer from "#/components/ui/footer.tsx"
 import { Header } from "#/components/ui/header.tsx"
+import { Prototype } from "#/components/ui/prototype/prototype.tsx"
 import { ShadowSinDevtools } from "#/integrations/tanstackDevtools/shadowSinDevtools.tsx"
 import { TanStackPacerDevtools } from "#/integrations/tanstackPacer/devtools.tsx"
 import { TanStackQueryDevtools } from "#/integrations/tanstackQuery/devtools.tsx"
@@ -26,28 +28,30 @@ export const Route = createRootRouteWithContext<RouterContext>()({
 function RootLayout() {
   return (
     <TanStackQueryProvider>
-      <Stack direction="column" sx={{ minHeight: "100vh", gap: 0 }}>
-        <Header />
+      <Prototype versions={dicePoolPrototypeVersions}>
+        <Stack direction="column" sx={{ minHeight: "100vh", gap: 0 }}>
+          <Header />
 
-        <Container
-          disableGutters
-          sx={{
-            maxWidth: 1200,
-            mx: "auto",
-            width: "100%",
-            flexGrow: 1,
-            display: "flex",
-            flexDirection: "column",
-          }}
-        >
-          <Outlet />
-        </Container>
+          <Container
+            disableGutters
+            sx={{
+              maxWidth: 1200,
+              mx: "auto",
+              width: "100%",
+              flexGrow: 1,
+              display: "flex",
+              flexDirection: "column",
+            }}
+          >
+            <Outlet />
+          </Container>
 
-        <Stack sx={{ padding: 1, gap: 1 }}>
-          <Divider />
-          <Footer />
+          <Stack sx={{ padding: 1, gap: 1 }}>
+            <Divider />
+            <Footer />
+          </Stack>
         </Stack>
-      </Stack>
+      </Prototype>
       <TanStackDevtools
         plugins={[
           ShadowSinDevtools,
